feat: start/stop all for ALL processes (aka home page)
