+++ src/Controllers/Home/Provider.php
@@ -15,9 +15,6 @@ class Provider implements ControllerProviderInterface
             );
 
             $controller->setTwig($app['twig']);
-            $controller->setRequest($app['request']);
-            $controller->setSession($app['session']);
-            $controller->setUrlGenerator($app['url_generator']);
 
             return $controller;
         };
@@ -40,44 +37,6 @@ class Provider implements ControllerProviderInterface
             ->assert('applicationName', '[\w-_.]+')
             ->method('GET')
             ->bind('applications');
-
-        $controllers
-            ->match('/servers/{serverName}/process/start/{processName}', 'controller.home:startProcessAction')
-            ->assert('serverName', '[\w-_.]+')
-            ->assert('processName', '[\w-_.]+')
-            ->method('GET')
-            ->bind('process.start');
-
-        $controllers
-            ->match('/servers/{serverName}/process/stop/{processName}', 'controller.home:stopProcessAction')
-            ->assert('serverName', '[\w-_.]+')
-            ->assert('processName', '[\w-_.]+')
-            ->method('GET')
-            ->bind('process.stop');
-
-        $controllers
-            ->match('/servers/{serverName}/process/startAll', 'controller.home:serverStartAllAction')
-            ->assert('serverName', '[\w-_.]+')
-            ->method('GET')
-            ->bind('process.server.startAll');
-
-        $controllers
-            ->match('/servers/{serverName}/process/stopAll', 'controller.home:serverStopAllAction')
-            ->assert('serverName', '[\w-_.]+')
-            ->method('GET')
-            ->bind('process.server.stopAll');
-
-        $controllers
-            ->match('/applications/{applicationName}/process/startAll', 'controller.home:applicationStartAllAction')
-            ->assert('applicationName', '[\w-_.]+')
-            ->method('GET')
-            ->bind('process.application.startAll');
-
-        $controllers
-            ->match('/applications/{applicationName}/process/stopAll', 'controller.home:applicationStopAllAction')
-            ->assert('applicationName', '[\w-_.]+')
-            ->method('GET')
-            ->bind('process.application.stopAll');
 
         return $controllers;
     }
